Add management proxy canister endpoint
