+++ packages/pocket-ic/Cargo.toml
@@ -56,3 +56,4 @@ ic-error-types = { path = "../../rs/types/error_types" }
 k256 = { workspace = true }
 lazy_static = { workspace = true }
 wat = { workspace = true }
+ic-management-canister-types = { path = "../../rs/types/management_canister_types" }
