+++ packages/pocket-ic/tests/tests.rs
@@ -1,11 +1,14 @@
 use candid::{decode_one, encode_one, CandidType, Decode, Deserialize, Encode, Principal};
 use ic_certification::Label;
+use ic_management_canister_types::NodeMetricsHistoryResponse;
 use ic_transport_types::Envelope;
 use ic_transport_types::EnvelopeContent::ReadState;
+use pocket_ic::common::rest::{BlockMakerConfigs, RawSubnetBlockmakerMetrics, TickConfigs};
 use pocket_ic::management_canister::{
     CanisterIdRecord, CanisterInstallMode, CanisterSettings, EcdsaPublicKeyResult,
-    HttpRequestResult, ProvisionalCreateCanisterWithCyclesArgs, SchnorrAlgorithm,
-    SchnorrPublicKeyArgsKeyId, SchnorrPublicKeyResult, SignWithBip341Aux, SignWithSchnorrAux,
+    HttpRequestResult, NodeMetricsHistoryArgs, ProvisionalCreateCanisterWithCyclesArgs,
+    SchnorrAlgorithm, SchnorrPublicKeyArgsKeyId, SchnorrPublicKeyResult, SignWithBip341Aux,
+    SignWithSchnorrAux,
 };
 use pocket_ic::{
     common::rest::{
@@ -2213,4 +2216,71 @@ fn test_reject_response_type() {
         // inspect message is always uncertified
         assert!(!err.certified);
     }
+}
+
+#[test]
+fn test_custom_blockmaker() {
+    // Setup PocketIC with initial subnets and records
+    let pocket_ic = PocketIcBuilder::new().with_application_subnet().build();
+
+    let topology = pocket_ic.topology();
+    let application_subnet = topology.get_app_subnets()[0];
+
+    // We create a canister on the app subnet.
+    let canister = pocket_ic.create_canister_on_subnet(None, None, application_subnet);
+    // We top up the canister with cycles and install the test canister WASM to them.
+    pocket_ic.add_cycles(canister, INIT_CYCLES);
+    pocket_ic.install_canister(canister, test_canister_wasm(), vec![], None);
+
+    let blockmaker_node = topology
+        .subnet_configs
+        .get(&application_subnet)
+        .unwrap()
+        .node_ids[0]
+        .clone();
+
+    let subnets_blockmakers = vec![RawSubnetBlockmakerMetrics {
+        subnet: application_subnet.into(),
+        blockmaker: blockmaker_node,
+        failed_blockmakers: vec![],
+    }];
+
+    let tick_configs = TickConfigs {
+        blockmakers: Some(BlockMakerConfigs {
+            blockmakers_per_subnet: subnets_blockmakers,
+        }),
+    };
+
+    pocket_ic.tick_with_configs(tick_configs.clone());
+    pocket_ic.tick_with_configs(tick_configs.clone());
+    pocket_ic.tick_with_configs(tick_configs.clone());
+    pocket_ic.tick_with_configs(tick_configs.clone());
+    pocket_ic.tick_with_configs(tick_configs.clone());
+
+    pocket_ic
+        //go to next day it should have recorder the metrics
+        .advance_time(std::time::Duration::from_secs(60 * 60 * 25));
+
+    pocket_ic.tick();
+    pocket_ic.tick();
+    pocket_ic.tick();
+    pocket_ic.tick();
+    pocket_ic.tick();
+
+    let response = pocket_ic
+        .query_call(
+            canister,
+            Principal::anonymous(),
+            "node_metrics_history_proxy",
+            Encode!(&NodeMetricsHistoryArgs {
+                subnet_id: application_subnet,
+                start_at_timestamp_nanos: 0,
+            })
+            .unwrap(),
+        )
+        .unwrap();
+
+    let res = Decode!(&response, Vec<NodeMetricsHistoryResponse>).unwrap();
+    println!("{:?}", res);
+    assert!(false);
 }
