+++ rs/types/management_canister_types/BUILD.bazel
@@ -9,6 +9,7 @@ permanent_whitelist = [
     "//rs/cycles_account_manager:__subpackages__",
     "//rs/messaging:__subpackages__",
     "//rs/management_canister_types/fuzz:__subpackages__",
+    "//packages/pocket-ic:__subpackages__",
 ]
 
 # These crates depend on this library directly for historical reasons
